+++ tests/testthat/test-verbs.R
@@ -19,49 +19,22 @@ test_that("plint works", {
   expect_s3_class(plint(x), "coo_single")
 })
 
-test_that("unpack works", {
-  expect_message(unpack("a"))
+test_that("unfold works", {
+  expect_message(unfold("a"))
 
-  x <- bot2$coo %>% unpack()
+  x <- bot2 %>% unfold()
   expect_false(is(x, "coo_tbl"))
   expect_false(is(x, "coo_list"))
   # also tests for 3 columns
   expect_true(all(colnames(x)[1:3] == c("shp", "x", "y")))
 
-  x <- bot2 %>% unpack
-  expect_false(is(x, "coo_tbl"))
-  expect_false(is(x, "coo_list"))
-  # also tests for 3 columns
-  expect_true(all(colnames(x)[1:3] == c("shp", "x", "y")))
-  # tests for names propagation
-  expect_true(is.character(x$shp))
-
-  # and that with NULL names, unpack still works
+  # and that with NULL names, unfold still works
   y <- bot2
   names(y$coo) <- NULL
-  x <- y %>% unpack
+  x <- y %>% unfold
   expect_false(is(x, "coo_tbl"))
   expect_false(is(x, "coo_list"))
   # also tests for 3 columns
   expect_true(all(c("shp", "x", "y") %in% colnames(x)))
   expect_true(is.numeric(x$shp))
 })
-
-# coe verbs ---------
-test_that("coe verbs work",{
-
-  x <- bot2 %>% efourier(4)
-  x$coe2 <- bot2$coo %>% efourier(6)
-
-  expect_equal(x %>% select_coe(coe) %>% only_coe() %>% colnames(), "coe")
-  expect_equal(x %>% select_coe(-coe) %>% only_coe() %>% colnames(), "coe2")
-
-  expect_equal(x %>% only_coe() %>% colnames(), c("coe", "coe2"))
-  expect_equal(x %>% drop_coe() %>% colnames(), bot2 %>% dplyr::select(-coo) %>% colnames())
-
-  expect_equal(x %>% ncoe, 2)
-  expect_equal(bot2 %>% ncoo, 1)
-})
-
-
-